requirements.txt
-i https://pypi.org/simple
aiofiles==0.3.2
bencode.py==2.0.0
blinker==1.4
click==6.7
git+https://github.com/5yncr/backend@master#egg=5yncr-backend
h11==0.7.0
h2==3.0.1
hpack==3.0.0
hyperframe==5.1.0
itsdangerous==0.24
jinja2==2.10
markupsafe==1.0
multidict==4.3.1
quart==0.5.0
sortedcontainers==2.0.1
wsproto==0.11.0
